Transcript Repair
This PR is for demonstrative purposes only!

This would merely make the transcript in 5.8 work as we expected at the time.

Pertex and I have "refurbished" most of this code in a different branch. In that branch:
- In desktop, the file saves as TXT rather than HTML
- The WebPlayer saves transcript data to localStorage
- We added numerous JS functions to tidy everything up
- Other small changes here and there

NOTE: Line 644 in playercore.js would throw an error every time text prints in a game created before Quest 5.8 **if** there is a JS variable named `savingTranscript` which is set to `true`, which is highly unlikely. I don't know how to get around that, regardless, without modifying a C# file to check the ASL version first.